+++ sample/src/main/java/io/snabble/testapp/BaseActivity.java
@@ -1,13 +1,8 @@
 package io.snabble.testapp;
 
-import android.app.Activity;
-import android.app.KeyguardManager;
-import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -17,21 +12,17 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import io.snabble.sdk.codes.ScannedCode;
-import io.snabble.sdk.ui.KeyguardHandler;
 import io.snabble.sdk.ui.SnabbleUI;
 import io.snabble.sdk.ui.SnabbleUICallback;
 import io.snabble.sdk.ui.integration.SelfScanningFragment;
 import io.snabble.sdk.ui.integration.ZebraSupport;
 import io.snabble.sdk.ui.scanner.ProductResolver;
-import io.snabble.sdk.ui.utils.KeyguardUtils;
 
 public abstract class BaseActivity extends AppCompatActivity implements SnabbleUICallback {
-    private static final int REQUEST_CODE_KEYGUARD = 0;
 
     private ProgressBar progressIndicator;
     private View content;
     private TextView sdkError;
-    private static KeyguardHandler keyguardHandler;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -156,30 +147,6 @@ public abstract class BaseActivity extends AppCompatActivity implements SnabbleU
     public void showPaymentCredentialsList() {
         Intent intent = new Intent(this, PaymentCredentialsListActivity.class);
         startActivity(intent);
-    }
-
-    @Override
-    public void requestKeyguard(KeyguardHandler keyguardHandler) {
-        KeyguardManager km = (KeyguardManager) getSystemService(Context.KEYGUARD_SERVICE);
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && KeyguardUtils.isDeviceSecure()) {
-            this.keyguardHandler = keyguardHandler;
-
-            Intent authIntent = km.createConfirmDeviceCredentialIntent(null, null);
-            startActivityForResult(authIntent, REQUEST_CODE_KEYGUARD);
-        } else {
-            keyguardHandler.onKeyguardResult(Activity.RESULT_OK);
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        if (requestCode == REQUEST_CODE_KEYGUARD) {
-            if(keyguardHandler != null) {
-                keyguardHandler.onKeyguardResult(resultCode);
-                keyguardHandler = null;
-            }
-        }
     }
 
     @Override