+++ ui/src/main/java/io/snabble/sdk/ui/KeyguardHandler.java
@@ -1,5 +1,0 @@
-package io.snabble.sdk.ui;
-
-public interface KeyguardHandler {
-    void onKeyguardResult(int resultCode);
-}