+++ ui/src/main/java/io/snabble/sdk/ui/SnabbleUICallback.java
@@ -1,7 +1,5 @@
 package io.snabble.sdk.ui;
 
-import android.content.Intent;
-
 public interface SnabbleUICallback {
     void showCheckout();
     void showScannerWithCode(String scannableCode);
@@ -10,18 +8,5 @@ public interface SnabbleUICallback {
     void showCreditCardInput();
     void showShoppingCart();
     void showPaymentCredentialsList();
-
-    /**
-     * Request a keyguard for user authentication. This will only be called if
-     * {@link io.snabble.sdk.UserPreferences#setRequireKeyguardAuthenticationForPayment(boolean)}
-     * is set to true.
-     *
-     * Typically you request a Keyguard using
-     * {@link android.app.KeyguardManager#createConfirmDeviceCredentialIntent(CharSequence, CharSequence)}.
-     *
-     * @param keyguardHandler A KeyguardHandler that needs to be supplied with the result of
-     *                        {@link android.app.Activity#onActivityResult(int, int, Intent)}.
-     */
-    void requestKeyguard(KeyguardHandler keyguardHandler);
     void goBack();
 }