+++ ui/src/main/java/io/snabble/sdk/ui/utils/KeyguardUtils.java
@@ -1,13 +1,10 @@
 package io.snabble.sdk.ui.utils;
 
-import android.app.Activity;
 import android.app.KeyguardManager;
 import android.content.Context;
 import android.os.Build;
 
 import io.snabble.sdk.Snabble;
-import io.snabble.sdk.ui.KeyguardHandler;
-import io.snabble.sdk.ui.SnabbleUI;
 
 public class KeyguardUtils {
     public static boolean isDeviceSecure() {
@@ -22,14 +19,6 @@ public class KeyguardUtils {
             return keyguardManager.isDeviceSecure(); // ignores SIM lock, API 23+
         } else {
             return keyguardManager.isKeyguardSecure();
-        }
-    }
-
-    public static void showKeyguard(KeyguardHandler keyguardHandler) {
-        if (Snabble.getInstance().getUserPreferences().isRequiringKeyguardAuthenticationForPayment()) {
-            SnabbleUI.getUiCallback().requestKeyguard(keyguardHandler);
-        } else {
-            keyguardHandler.onKeyguardResult(Activity.RESULT_OK);
         }
     }
 }